+++ apps/perms/api/user_permission/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-from .tree import *  # noqa fix circular import
 from .nodes import *
 from .assets import *
 from .accounts import *
+from .tree import *
